Move error types into error.rs
