+++ raylib/src/core/models.rs
@@ -3,65 +3,9 @@
 use crate::core::math::{BoundingBox, Vector3};
 use crate::core::texture::Image;
 use crate::core::{RaylibHandle, RaylibThread};
-use crate::{consts, ffi};
+use crate::{consts, ffi, error::{RaylibLoadMaterialError, RaylibLoadModelAnimError, RaylibLoadModelError, RaylibSetMaterialError}};
 use std::ffi::CString;
 use std::os::raw::c_void;
-
-#[derive(Debug)]
-pub enum RaylibLoadModelError<'a> {
-    LoadFromFileFailed(&'a str),
-    LoadFromMeshFailed,
-}
-impl std::fmt::Display for RaylibLoadModelError<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::LoadFromFileFailed(path) => write!(f, "could not load model\npath: {path:?}"),
-            Self::LoadFromMeshFailed => f.write_str("could not load model from mesh"),
-        }
-    }
-}
-impl std::error::Error for RaylibLoadModelError<'_> {}
-
-#[derive(Debug)]
-pub enum RaylibLoadModelAnimError<'a> {
-    NoAnimationsLoaded(&'a str),
-}
-impl std::fmt::Display for RaylibLoadModelAnimError<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::NoAnimationsLoaded(path) => write!(f, "no model animations loaded\npath: {path:?}"),
-        }
-    }
-}
-impl std::error::Error for RaylibLoadModelAnimError<'_> {}
-
-#[derive(Debug)]
-pub enum RaylibSetMaterialError {
-    MeshIdOutOfBounds,
-    MaterialIdOutOfBounds,
-}
-impl std::fmt::Display for RaylibSetMaterialError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::MeshIdOutOfBounds => f.write_str("mesh_id greater than mesh count"),
-            Self::MaterialIdOutOfBounds => f.write_str("material_id greater than material count"),
-        }
-    }
-}
-impl std::error::Error for RaylibSetMaterialError {}
-
-#[derive(Debug)]
-pub enum RaylibLoadMaterialError<'a> {
-    NoneLoaded(&'a str),
-}
-impl std::fmt::Display for RaylibLoadMaterialError<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::NoneLoaded(path) => write!(f, "no materials loaded\npath: {path:?}"),
-        }
-    }
-}
-impl std::error::Error for RaylibLoadMaterialError<'_> {}
 
 fn no_drop<T>(_thing: T) {}
 make_thin_wrapper!(Model, ffi::Model, ffi::UnloadModel);
@@ -115,7 +59,7 @@ impl RaylibHandle {
         let m = unsafe { ffi::LoadModel(c_filename.as_ptr()) };
         if m.meshes.is_null() && m.materials.is_null() && m.bones.is_null() && m.bindPose.is_null()
         {
-            return Err(RaylibLoadModelError::LoadFromFileFailed(filename));
+            return Err(RaylibLoadModelError::LoadFromFileFailed { filename });
         }
         // TODO check if null pointer checks are necessary.
         Ok(Model(m))
@@ -145,7 +89,7 @@ impl RaylibHandle {
         let mut m_size = 0;
         let m_ptr = unsafe { ffi::LoadModelAnimations(c_filename.as_ptr(), &mut m_size) };
         if m_size <= 0 {
-            return Err(RaylibLoadModelAnimError::NoAnimationsLoaded(filename));
+            return Err(RaylibLoadModelAnimError::NoAnimationsLoaded { filename });
         }
         let mut m_vec = Vec::with_capacity(m_size as usize);
         for i in 0..m_size {
@@ -528,7 +472,7 @@ impl Material {
         let mut m_size = 0;
         let m_ptr = unsafe { ffi::LoadMaterials(c_filename.as_ptr(), &mut m_size) };
         if m_size <= 0 {
-            return Err(RaylibLoadMaterialError::NoneLoaded(filename));
+            return Err(RaylibLoadMaterialError::NoneLoaded { filename });
         }
         let mut m_vec = Vec::with_capacity(m_size as usize);
         for i in 0..m_size {
